Добавлен нестатический вариант MachineConfig. Переименовал проект QSMachineConfig в QS.MachineConfig
