--- QSMachineConfig/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="NLog" version="4.7.5" targetFramework="net40" />
-</packages>
